refactor(pulumi): move provider specific logic `provider.ts`
