+++ plugins/pulumi/handlers.ts
@@ -6,8 +6,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import {pulumi} from "./cli"
-import {ProviderHandlers} from "@garden-io/sdk/types"
+import { pulumi } from "./cli"
+import { ProviderHandlers } from "@garden-io/sdk/types"
 import {
   applyConfig,
   clearStackVersionTag,
@@ -20,11 +20,12 @@ import {
   selectStack,
   setStackVersionTag,
 } from "./helpers"
-import {PulumiDeploy, PulumiProvider} from "./config"
+import { PulumiDeploy } from "./config"
+import { PulumiProvider } from "./provider"
 import chalk from "chalk"
-import {DeployActionHandlers} from "@garden-io/core/build/src/plugin/action-types"
-import {DeployState} from "@garden-io/core/build/src/types/service"
-import {deployStateToActionState} from "@garden-io/core/build/src/plugin/handlers/Deploy/get-status"
+import { DeployActionHandlers } from "@garden-io/core/build/src/plugin/action-types"
+import { DeployState } from "@garden-io/core/build/src/types/service"
+import { deployStateToActionState } from "@garden-io/core/build/src/plugin/handlers/Deploy/get-status"
 
 export const cleanupEnvironment: ProviderHandlers["cleanupEnvironment"] = async (_params) => {
   // To properly implement this handler, we'd need access to the config graph (or at least the list of pulumi services
