+++ plugins/pulumi/helpers.ts
@@ -21,7 +21,8 @@ import { loadAndValidateYaml } from "@garden-io/core/build/src/config/base"
 import { getPluginOutputsPath } from "@garden-io/sdk"
 import { Log, PluginContext } from "@garden-io/sdk/types"
 import { defaultPulumiEnv, pulumi } from "./cli"
-import { PulumiDeploy, PulumiProvider } from "./config"
+import { PulumiDeploy } from "./config"
+import { PulumiProvider } from "./provider"
 import { deline } from "@garden-io/sdk/util/string"
 import { Resolved } from "@garden-io/core/build/src/actions/types"
 import { ActionLog } from "@garden-io/core/build/src/logger/log-entry"
