+++ plugins/pulumi/index.ts
@@ -11,7 +11,8 @@ import { dedent } from "@garden-io/sdk/util/string"
 import { deletePulumiDeploy, deployPulumi, getPulumiDeployStatus } from "./handlers"
 import { getPulumiCommands } from "./commands"
 import { pulumiCliSPecs } from "./cli"
-import { PulumiDeployConfig, pulumiDeployOutputsSchema, pulumiDeploySchema, pulumiProviderConfigSchema } from "./config"
+import { PulumiDeployConfig, pulumiDeployOutputsSchema, pulumiDeploySchema } from "./config"
+import { pulumiProviderConfigSchema } from "./provider"
 import { ExecBuildConfig } from "@garden-io/core/build/src/plugins/exec/build"
 import { join } from "path"
 import { pathExists } from "fs-extra"
