+++ plugins/pulumi/test/handlers.ts
@@ -13,7 +13,7 @@ import execa from "execa"
 import { pathExists } from "fs-extra"
 import { join, resolve } from "path"
 import { deployPulumi, getPulumiDeployStatus } from "../handlers"
-import { PulumiProvider } from "../config"
+import { PulumiProvider } from "../provider"
 import { gardenPlugin as pulumiPlugin } from ".."
 import { expect } from "chai"
 import { getStackVersionTag } from "../helpers"
